requirements.txt
flask==3.0.0
google-cloud-firestore==2.14.0
gunicorn==21.2.0